Fix threading issues on UNIX
Add pthread flag for UNIX builds of the test binary.

The flag is required because the tests currently do not make use of
threading. As a result, the compiler replaces lock calls with noop.

At a later stage, the tests should be updated so that they actually make
use of threading. Then, it might be possible to remove the flag.